services/media_ingestion/requirements.txt
fastapi
uvicorn
yt-dlp
ffmpeg-python
requests
psycopg2-binary
